threading fix (maybe incomplete)

+++ bitmap/float_bm4.cpp
@@ -40,7 +40,7 @@ struct SSBumpCalculationContext								// what each thread needs to see
 };
 
 
-static unsigned SSBumpCalculationThreadFN(void* ctx1)
+static unsigned long long SSBumpCalculationThreadFN(void* ctx1)
 {
 	SSBumpCalculationContext* ctx = (SSBumpCalculationContext*)ctx1;
