+++ gameui/BasePanel.cpp
@@ -2512,7 +2512,7 @@ bool CBasePanel::IsPromptableCommand( const char *command )
 //-------------------------
 // Purpose: Job wrapper
 //-------------------------
-static unsigned PanelJobWrapperFn( void *pvContext )
+static unsigned long long PanelJobWrapperFn( void *pvContext )
 {
 	CBasePanel::CAsyncJobContext *pAsync = reinterpret_cast< CBasePanel::CAsyncJobContext * >( pvContext );
